+++ src/generic-table/interfaces/gt-options.ts
@@ -1,6 +1,8 @@
 export interface GtOptions {
-  cache: boolean;
-  debounceTime: number;
-  csvDelimiter: string;
-  stack: boolean;
+  cache?: boolean;
+  debounceTime?: number;
+  csvDelimiter?: string;
+  stack?: boolean;
+  lazyLoad?: boolean;
+  highlightSearch?: boolean;
 }